% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arr_last_unique}
\alias{arr_last_unique}
\title{Determine the last index containing a unique value}
\usage{
arr_last_unique(mat, dim_index)
}
\description{
Determine the last index containing a unique value
}
