% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{as.character.heRoFormula}
\alias{as.character.heRoFormula}
\title{Convert a Formula to Character}
\usage{
\method{as.character}{heRoFormula}(x, ...)
}
\arguments{
\item{x}{A heRoFormula object.}

\item{...}{additional unused arguments.}
}
\value{
An atomic character vector representing the R-expression.
}
\description{
Takes a heRoFormula object and converts it to its character representation.
}
