% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{as.heRoFormula}
\alias{as.heRoFormula}
\title{Convert to Formula}
\usage{
as.heRoFormula(x)
}
\arguments{
\item{x}{Object to be converted to heRovar.}
}
\value{
A heRovar object.
}
\description{
Convert an object to a heRovar object.
}
\keyword{internal}
