% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as.heRovar}
\alias{as.heRovar}
\title{Convert to Variable}
\usage{
as.heRovar(x)
}
\arguments{
\item{x}{Object to be converted to heRovar.}
}
\value{
A heRovar object.
}
\description{
Convert an object to a heRovar object.
}
\keyword{internal}
