% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap sampling of a data frame}
\usage{
bootstrap(x, id = NULL, strata = NULL, weight = NULL)
}
\arguments{
\item{x}{The data frame to be sampled}

\item{id}{An optional column name representing the unique identifier. If
not specified, each row is assumed to represent a unique observation.}

\item{strata}{An optional character vector of column names representing the
strata within which resampling should be performed.}

\item{weight}{An optional column name representing the probability weight for each
observation.}
}
\description{
The `bootstrap` function can be used to handle sampling
of tables of individual patient data by selecting observations
from the table with replacement
}
