% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_markov_lf.R
\name{calc_compl_probs}
\alias{calc_compl_probs}
\title{Calculate complementary probabilities in an evaluated transition matrix}
\usage{
calc_compl_probs(mat)
}
\description{
Calculate complementary probabilities in an evaluated transition matrix
}
