% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Perform Error Checking}
\usage{
check(x, ...)
}
\arguments{
\item{x}{The object to be checked}

\item{...}{Additional arguments to be handled
by the method}
}
\description{
Performs error checking on a given model
object.
}
