% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_tbl}
\alias{check_tbl}
\title{Check Table}
\usage{
check_tbl(df, spec, context)
}
\arguments{
\item{df}{The input dataframe to be checked}

\item{spec}{The specification dataframe}

\item{context}{A string used in error messages to indicate the type of input dataframe}
}
\value{
The input dataframe with missing values imputed
}
\description{
Check a model inputs dataframe based on a given specification dataframe. Specification
dataframes are used to check input dataframes, ensure that required columns are present
and impute missing values for non-required columns.
}
