% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{clone_namespace}
\alias{clone_namespace}
\title{Clone a Namespace}
\usage{
clone_namespace(x)
}
\arguments{
\item{x}{The namespace to be cloned.}
}
\value{
An identical copy of the namespace.
}
\description{
Clones a namespace object.
}
\keyword{internal}
