% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{define_parameter}
\alias{define_parameter}
\title{Capture a Parameter}
\usage{
define_parameter(string)
}
\arguments{
\item{text}{a string containing an R-expression}
}
\value{
A parameter object
}
\description{
Takes a string and converts into it a parameter object
}
