% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_parameters}
\alias{define_parameters}
\alias{define_parameters_tabular}
\alias{define_parameters_}
\title{Define Model Parameters}
\usage{
define_parameters(df)

define_parameters_tabular(x)

define_parameters_(.dots)
}
\arguments{
\item{x}{A data frame with columns `name` and `formula`}

\item{.dots}{An object of type `lazy_dots` representing
the unevaluated parameters.}

\item{...}{A set of named parameters}
}
\description{
Define a set of named R-expressions representing
model parameters.
}
