% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_variable}
\alias{define_variable}
\title{Capture a Variable}
\usage{
define_variable(string)
}
\arguments{
\item{text}{An atomic character vector containing an R-expression.}
}
\value{
A heRovar object representing the given R-expression.
}
\description{
Takes a string representing an R-expression and turns it into a heRovar
object. If the resulting R-expression is invalid, it will be replaced
with an expression that evaluates to an error string.
}
\keyword{internal}
