% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_variable_list}
\alias{define_variable_list}
\title{Capture a List of Variables}
\usage{
define_variable_list(df)
}
\arguments{
\item{df}{A data.frame with columns for `name` and `formula`.}
}
\value{
A heRovar_list object representing the given variables.
}
\description{
Takes a data.frame representing a set of heRo variables and converts it to
a heRovar_list object.
}
\keyword{internal}
