% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{err_name_string}
\alias{err_name_string}
\title{Generate List of Names for Error Messages}
\usage{
err_name_string(x)
}
\arguments{
\item{x}{character vector of names}
}
\value{
character with comma-separated list of quoted names
}
\description{
Generates a comma-separated character of quoted names in order to list
items referenced in error messages.
}
