% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{eval_parameter}
\alias{eval_parameter}
\title{Evaluate a Single Parameter}
\usage{
eval_parameter(x, df)
}
\arguments{
\item{x}{An object of class `lazy`}

\item{df}{A `data.frame` representing existing parameter values}
}
\description{
Evaluates a single parameter in the context of the given
`data.frame` of previously evaluated parameters.
}
\keyword{internal}
