% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{eval_parameters}
\alias{eval_parameters}
\title{Evaluate Parameters}
\usage{
eval_parameters(x, df)
}
\arguments{
\item{x}{An object of class `param_uneval`}

\item{df}{A `data.frame` representing existing parameter values}
}
\description{
Evaluates a parameters object and returns a table with
the values of each parameter in each cycle.
}
\keyword{internal}
