% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.r
\name{eval_trans_markov_lf}
\alias{eval_trans_markov_lf}
\title{Evaluate a Longform Transition Matrix}
\usage{
eval_trans_markov_lf(df, ns, simplify = FALSE)
}
\description{
Evaluate a Longform Transition Matrix
}
