% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{evaluate.heRovar}
\alias{evaluate.heRovar}
\title{Evaluates a variable}
\usage{
\method{evaluate}{heRovar}(x, ns, ...)
}
\arguments{
\item{x}{A heRovar object.}
}
\value{
An atomic character vector representing the R-expression.
}
\description{
Takes a heRovar object and converts it to its character representation.
}
