% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Model Parameters}
\usage{
evaluate(x, ns, ...)
}
\arguments{
\item{x}{The object to be evaluated}

\item{ns}{The namespace object in which to evaluate}

\item{...}{Optional arguments to be handled by method}
}
\description{
Evaluates an object containing one or more
R-expressions representing model calculations.
}
