% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a Model}
\usage{
evaluate_model(model, ...)
}
\arguments{
\item{model}{A heRo_model object.}
}
\value{
A list containing the results of the model.
}
\description{
Takes a model specification object and runs the model.
}
