% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_model_segment}
\alias{evaluate_model_segment}
\title{Evaluate a Model Segment}
\usage{
evaluate_model_segment(segment, model, env, ...)
}
\arguments{
\item{x}{A heRovar object.}

\item{ns}{A namespace object containing evaluated variables that the heRovar
object may reference.}
}
\value{
The result of evaluating the heRovar object.
}
\description{
Evaluates a segment of a health economic model. A segment is defined as
a single run of the model for a given strategy & group. Explicitly defining
models in terms of these segments makes it possible to parallelize model
calculations over strategies, groups, and simulations.
}
\keyword{internal}
