% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longform_matrix.R
\name{evaluate_trans_row}
\alias{evaluate_trans_row}
\title{Evaluate a row of a longform transition matrix}
\usage{
evaluate_trans_row(row, ns, simplify = F)
}
\description{
Evaluate a row of a longform transition matrix
}
