% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_variable}
\alias{evaluate_variable}
\title{Evaluates a variable}
\usage{
evaluate_variable(x, ns, ...)
}
\arguments{
\item{x}{A heRovar object.}

\item{ns}{A namespace object containing evaluated variables that the heRovar
object may reference.}
}
\value{
The result of evaluating the heRovar object.
}
\description{
Takes a heRovar object and converts it to its character representation.
}
\keyword{internal}
