% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{export}
\alias{export}
\title{Export a Variable}
\usage{
export(x)
}
\arguments{
\item{x}{The object being exported.}
}
\value{
a list containing the class of the result, and the output of its
print and summary methods.
}
\description{
Generates output representing the evaluated value for a given variable. This
is done by outputing its class, and the result of its print and summary
methods.
}
