% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_faslsy_chr}
\alias{is_faslsy_chr}
\title{Check If Character is Empty/Missing}
\usage{
is_faslsy_chr(x)
}
\arguments{
\item{x}{character vector to be checked}
}
\value{
logical vector indicating whether the elements of `x` are empty or missing.
}
\description{
Check if a character is either `NA` or empty.
}
