% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_valid_name}
\alias{is_valid_name}
\title{Check if Name is Valid}
\usage{
is_valid_name(x)
}
\arguments{
\item{x}{character vector of names to check}
}
\value{
logical vector indicating whether the elements of `x` are valid.
}
\description{
Checks character vector to see if elements follow the rules of heRomod2 variable names,
which must start with a letter and include only letters, numbers, and underscores.
}
