% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{lognormal}
\alias{lognormal}
\title{Lognormal distribution}
\usage{
lognormal(mean, sd, meanlog, sdlog)
}
\arguments{
\item{mean}{Mean of parameter}

\item{sd}{Standard deviation of parameter}

\item{meanlog}{Mean of `log(parameter)`}

\item{sdlog}{Standard deviation of `log(parameter)`}
}
\description{
The `lognormal` function can be used to sample a parameter
whose natural log is distributed normally.  Lognormal distribution
can be specified by providing the mean and standard deviation of the
parameter either on the natural or log scale by supplying the
appropriate arguments.
}
\author{
Antoine Filipovic-Pierucci
}
