% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_specs.R
\docType{data}
\name{model_input_specs}
\alias{model_input_specs}
\title{Input Data Specifications}
\format{A list of dataframes}
\usage{
model_input_specs
}
\description{
A list of dataframes describing the format of model input dataframes,
including which columns are required and how missing values should be
imputed.
}
\keyword{datasets}
