% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{normal}
\alias{normal}
\title{Normal distribution}
\usage{
normal(mean, sd)
}
\arguments{
\item{mean}{Mean of parameter}

\item{sd}{Standard deviation of parameter}
}
\description{
The `normal` function can be used to sample a normally-distributed
parameter.
}
\author{
Antoine Filipovic-Pierucci
}
