% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openqaly-package.R
\docType{package}
\name{openqaly-package}
\alias{openqaly}
\alias{openqaly-package}
\title{openqaly: Reproducible Health Economic Modeling}
\description{
An R package for creating and running reproducible and transparent health economic models. Models of various types are creating using a consistent tabular structure which defines features of the model such as the strategies being compared and the health states between which patients transit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://PolicyAnalysisInc.github.io/openqaly/}
  \item \url{https://github.com/PolicyAnalysisInc/openqaly}
  \item Report bugs at \url{https://github.com/PolicyAnalysisInc/openqaly/issues}
}

}
\author{
\strong{Maintainer}: Jordan Amdahl \email{jordan.amdahl@avalerehealth.com}

Other contributors:
\itemize{
  \item Avalere Health [copyright holder]
}

}
\keyword{internal}
