% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{print.heRovar}
\alias{print.heRovar}
\title{Convert Variables Object to Character}
\usage{
\method{print}{heRovar}(x, ...)
}
\arguments{
\item{x}{A heRovar object.}
}
\value{
An atomic character vector representing the R-expression.
}
\description{
Takes a heRovar object and converts it to its character representation.
}
