% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_workbook}
\alias{read_workbook}
\title{Read an Excel Workbook}
\usage{
read_workbook(path)
}
\arguments{
\item{path}{the path to the workbook.}
}
\value{
a named list of data.frames
}
\description{
Takes the path of an excel workbook and reads it in as a named list of
data frames.
}
