% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{resample}
\alias{resample}
\title{Resample Model Parameters}
\usage{
resample(model, n, segments, corr = NULL)
}
\arguments{
\item{model}{A heRomodel object}

\item{n}{the number of simulations}

\item{segments}{the segments for which resampling will be done}

\item{corr}{an optional correlation matrix}
}
\value{
a data.frame with resampled data by segment
}
\description{
Resamples the parameters for a model based on the sampling specifications.
}
