$% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Run a Model}
\usage{
run_model(model, ...)
}
\arguments{
\item{model}{A heRo_model object.}

\item{...}{additional arguments.}
}
\value{
A list containing the results of the model.
}
\description{
Takes a model specification object and runs the model.
}
