% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort.heRovar_list}
\alias{sort.heRovar_list}
\title{Sort Variables List}
\usage{
\method{sort}{heRovar_list}(x, ...)
}
\arguments{
\item{x}{The heRovar_list object to be sorted}

\item{...}{Unused variables to match sort call signature}
}
\description{
Sorts a heRovar_list object in order to solve issues of dependency
resolution and identify any circular references.
}
