% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort.parameters}
\alias{sort.parameters}
\title{Sort Model Parameters}
\usage{
\method{sort}{parameters}(x, ...)
}
\arguments{
\item{x}{The parameters object to be sorted}

\item{...}{Unused parameters to match sort call
signature}
}
\description{
Sorts a parameters object in order to solve issues
of dependency-resolution and identify any circular
references.
}
