% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{sort_parameters}
\alias{sort_parameters}
\title{Sort Parameters}
\usage{
sort_parameters(x)
}
\arguments{
\item{x}{An object of class `param_uneval`}
}
\value{
A sorted object of class `param_uneval`
}
\description{
Sorts unevaluated parameters such that no parameter is
referenced before it is defined.  Will detect and throw
an error in case of circular references.
}
\keyword{internals}
