% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vswitch}
\alias{vswitch}
\title{Vectorized Switch Statement}
\usage{
vswitch(x, ...)
}
\arguments{
\item{x}{The condition statement}

\item{...}{name-value pairs where name repesents cases of condition
statement and values represent the value it will take on in each case.}
}
\value{
Returns value for the given case.
}
\description{
Vectorized Switch Statement
}
