% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as.character.heRovar}
\alias{as.character.heRovar}
\title{Convert a Variable to Character}
\usage{
\method{as.character}{heRovar}(x, ...)
}
\arguments{
\item{x}{A heRovar object.}

\item{...}{additional unused arguments.}
}
\value{
An atomic character vector representing the R-expression.
}
\description{
Takes a heRovar object and converts it to its character representation.
}
