% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{as.heRovar}
\alias{as.heRovar}
\title{Convert to Variable}
\usage{
as.heRovar(x)
}
\arguments{
\item{x}{Object to be converted to heRovar.}
}
\value{
A heRovar object.
}
\description{
Takes a list containing the elements of a heRovar object and returns a
heRovar object.
}
