% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{define_decision_trees}
\alias{define_decision_trees}
\title{Define a decision Tree}
\usage{
define_decision_trees(trees)
}
\arguments{
\item{trees}{a data.frame representing the decision trees.}
}
\description{
Define a decision Tree
}
