% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{define_namespace}
\alias{define_namespace}
\title{Define a Namespace Object}
\usage{
define_namespace(df, parent = new.env())
}
\arguments{
\item{df}{A data frame of pre-existing parameter values}

\item{parent}{An optional environment to serve as the
parent to the namespace environment}
}
\description{
Creates a new namespace object which can be used
to store evaluated parameters.  A namespace object
combines a data frame for storing vector parameters
with an environment for storing non-vector parameters.
}
